+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Snap, create: Fixed possible deadlock
 - Incorrect resource definition already exists error message
+- sos-report: Controller not always dumping full log content
 
 ## [1.28.0] - 2024-07-11